refactor(auth): move auth tokens into private module
