+++ tests/unit/test_public_shims.py
@@ -114,6 +114,7 @@ def test_public_import_manifest_has_no_duplicates() -> None:
 def test_public_facade_imports_are_identity_reexports() -> None:
     """Compatibility facades must keep returning the canonical public objects."""
     import notebooklm
+    import notebooklm._auth.tokens as private_tokens
     import notebooklm.auth as public_auth
     import notebooklm.rpc as public_rpc
     import notebooklm.rpc.overrides as rpc_overrides
@@ -121,6 +122,7 @@ def test_public_facade_imports_are_identity_reexports() -> None:
     import notebooklm.types as public_types
 
     assert notebooklm.AuthTokens is public_auth.AuthTokens
+    assert public_auth.AuthTokens is private_tokens.AuthTokens
     assert notebooklm.ConnectionLimits is public_types.ConnectionLimits
     assert public_rpc.RPCMethod is rpc_types.RPCMethod
     assert public_rpc.resolve_rpc_id is rpc_overrides.resolve_rpc_id
